Move tests to /test folder
Revert to Robolectric 4.3 (4.3.1 requires Java 9, which requires manual configuration)
