+++ library/build.gradle
@@ -28,10 +28,10 @@ android {
         targetSdkVersion 29
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         consumerProguardFiles 'consumer-rules.pro'
         // This enables us to tell when we're running unit tests on Travis (#573)
         buildConfigField("String", "TRAVIS", "\"" + System.getenv('TRAVIS') + "\"")
+        multiDexEnabled true
     }
     buildTypes {
         release {
@@ -45,6 +45,16 @@ android {
         timeOutInMs 10 * 60 * 1000  // 10 minutes
         installOptions "-d", "-t"
     }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+    // Robolectric
+//    sourceSets {
+//        test {
+//            res.srcDirs = ['src/test/res', 'src/main/res']
+//        }
+//    }
     testOptions {
         unitTests {
             includeAndroidResources = true
@@ -58,7 +68,22 @@ dependencies {
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'androidx.test.ext:junit:1.1.1'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
-    androidTestImplementation 'org.robolectric:robolectric:4.3.1'
+    testImplementation ('org.robolectric:robolectric:4.3') {
+        exclude group: 'org.apache.maven', module: 'maven-artifact'
+        exclude group: 'org.apache.maven', module: 'maven-artifact-manager'
+        exclude group: 'org.apache.maven', module: 'maven-model'
+        exclude group: 'org.apache.maven', module: 'maven-plugin-registry'
+        exclude group: 'org.apache.maven', module: 'maven-profile'
+        exclude group: 'org.apache.maven', module: 'maven-project'
+        exclude group: 'org.apache.maven', module: 'maven-settings'
+        exclude group: 'org.apache.maven', module: 'maven-error-diagnostics'
+        exclude group: 'org.apache.maven.wagon'
+        exclude group: 'org.apache.maven', module: 'maven-ant-tasks'
+        exclude group: 'org.codehaus', module: 'classworlds'
+        exclude group: 'org.codehaus', module: 'plexus'
+    }
+    // Required to support multidex for minSdkVersion <= 20
+    implementation 'androidx.multidex:multidex:2.0.1'
 }
 
 jacoco {
