+++ library/src/test/java/com/google/maps/android/data/geojson/GeoJsonPointStyleTest.java
@@ -15,8 +15,6 @@
  */
 package com.google.maps.android.data.geojson;
 
-import androidx.test.platform.app.InstrumentationRegistry;
-
 import com.google.android.gms.maps.MapsInitializer;
 import com.google.android.gms.maps.model.BitmapDescriptor;
 import com.google.android.gms.maps.model.BitmapDescriptorFactory;
@@ -29,6 +27,8 @@ import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
 
 import java.util.Arrays;
+
+import androidx.test.platform.app.InstrumentationRegistry;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
