+++ library/src/test/java/com/google/maps/android/data/kml/KmlFeatureParserTest.java
@@ -15,21 +15,18 @@
  */
 package com.google.maps.android.data.kml;
 
-import androidx.test.platform.app.InstrumentationRegistry;
-
+import com.google.maps.android.R;
 import com.google.maps.android.data.Geometry;
-import com.google.maps.android.test.R;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
 import org.xmlpull.v1.XmlPullParser;
-import org.xmlpull.v1.XmlPullParserFactory;
 
-import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.List;
 
+import static com.google.maps.android.data.kml.KmlTestUtil.createParser;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
@@ -37,19 +34,6 @@ import static org.junit.Assert.assertTrue;
 
 @RunWith(RobolectricTestRunner.class)
 public class KmlFeatureParserTest {
-    private XmlPullParser createParser(int res) throws Exception {
-        InputStream stream =
-                InstrumentationRegistry.getInstrumentation()
-                        .getTargetContext()
-                        .getResources()
-                        .openRawResource(res);
-        XmlPullParserFactory factory = XmlPullParserFactory.newInstance();
-        factory.setNamespaceAware(true);
-        XmlPullParser parser = factory.newPullParser();
-        parser.setInput(stream, null);
-        parser.next();
-        return parser;
-    }
 
     @Test
     public void testPolygon() throws Exception {
