+++ library/src/test/java/com/google/maps/android/heatmaps/GradientTest.java
@@ -16,12 +16,12 @@
 
 package com.google.maps.android.heatmaps;
 
-import android.graphics.Color;
-import android.os.Build;
-
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
+
+import android.graphics.Color;
+import android.os.Build;
 
 import static android.graphics.Color.BLUE;
 import static android.graphics.Color.GREEN;
