+++ library/src/test/java/com/google/maps/android/heatmaps/UtilTest.java
