+++ library/src/test/res/raw/amu_ballon_gx_prefix.kml
