+++ library/src/test/res/raw/amu_basic_folder.kml
