+++ library/src/test/res/raw/amu_basic_placemark.kml
