+++ library/src/test/res/raw/amu_cdata.kml
