+++ library/src/test/res/raw/amu_default_balloon.kml
