+++ library/src/test/res/raw/amu_document_nest.kml
