+++ library/src/test/res/raw/amu_draw_order_ground_overlay.kml
