+++ library/src/test/res/raw/amu_empty_hotspot.kml
