+++ library/src/test/res/raw/amu_extended_data.kml
