+++ library/src/test/res/raw/amu_ground_overlay.kml
