+++ library/src/test/res/raw/amu_ground_overlay_color.kml
