+++ library/src/test/res/raw/amu_inline_style.kml
