+++ library/src/test/res/raw/amu_multigeometry_placemarks.kml
