+++ library/src/test/res/raw/amu_multiple_placemarks.kml
