+++ library/src/test/res/raw/amu_nested_folders.kml
