+++ library/src/test/res/raw/amu_poly_style_boolean_alpha.kml
