+++ library/src/test/res/raw/amu_unknown_folder.kml
