+++ library/src/test/res/raw/amu_unsupported.kml
