+++ library/src/test/res/raw/amu_visibility_ground_overlay.kml
